third_party/WebKit/LayoutTests/fast/serviceworker/serviceworkercontainer-interface.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
function getPropertyDescriptor(name) {
  return (
      Object.getOwnPropertyDescriptor(ServiceWorkerContainer.prototype, name) ||
      // FIXME: Remove this when http://crbug.com/43394 is fixed.
      Object.getOwnPropertyDescriptor(navigator.serviceWorker, name));
}

test(function() {
  assert_true('ServiceWorkerContainer' in window,
              'the constructor should be defined');
  assert_throws(new TypeError, function() { new ServiceWorkerContainer(); },
                'the constructor should not be callable with "new"');
  assert_throws(new TypeError, function() { ServiceWorkerContainer(); },
                'the constructor should not be callable');
}, 'ServiceWorkerContainer interface');
</script>
